+++ selfdrive/car/modules/qt/tsettings.cc
@@ -85,6 +85,12 @@ TinklaTogglesPanel::TinklaTogglesPanel(SettingsWindow *parent) : ListWidget(pare
       "../assets/offroad/icon_settings.png",
       "","","",0.0,0.0,0.0,0.0, TINKLA_TOGGLE
       },
+    {"TinklaFlipScreen",
+      "Flip screen (for RHD)",
+      "Flips the screen for people driving on the wrong side of the road. Changes to C2 hardware are needed. C2 only.",
+      "../assets/offroad/icon_settings.png",
+      "","","",0.0,0.0,0.0,0.0, TINKLA_TOGGLE
+      },
     {"TinklaShutdownAfter",
       "Shutdown after # of hours",
       "Shutdown device after number of hours when car is off",
@@ -260,6 +266,13 @@ TeslaPreApTogglesPanel::TeslaPreApTogglesPanel(SettingsWindow *parent) : ListWid
     QProcess::startDetached("/data/openpilot/selfdrive/car/modules/radarFlasher/flashTeslaRadar");
   });
 
+  QPushButton *ivs_flash_btn = new QPushButton("Flash Vacuum Sens");
+  vin_radar_btn->setObjectName("ivs_flash_btn");
+
+  QObject::connect(ivs_flash_btn, &QPushButton::clicked, [=](){
+    QProcess::startDetached("/data/openpilot/panda/board/ibooster_vacuum_sensor/flashIvs");
+  });
+
   setStyleSheet(R"(
     #flash_btn { height: 120px; border-radius: 15px; background-color: #393939; }
     #flash_btn:pressed { background-color: #4a4a4a; }
@@ -267,10 +280,13 @@ TeslaPreApTogglesPanel::TeslaPreApTogglesPanel(SettingsWindow *parent) : ListWid
     #flash_pedal_btn:pressed { background-color: #4a4a4a; }
     #vin_radar_btn { height: 120px; border-radius: 15px; background-color: #393939; }
     #vin_radar_btn:pressed { background-color: #4a4a4a; }
+    #ivs_flash_btn { height: 120px; border-radius: 15px; background-color: #393939; }
+    #ivs_flash_btn:pressed { background-color: #4a4a4a; }
   )");
   addItem(flash_btn);
   addItem(flash_pedal_btn);
   addItem(vin_radar_btn);
+  addItem(ivs_flash_btn);
 }
 
 TeslaTogglesPanel::TeslaTogglesPanel(SettingsWindow *parent) : ListWidget(parent) {